+++ telemetry/telemetry.gradle
@@ -22,6 +22,7 @@ ext {
     'datadog.telemetry.Uname.UtsNameLinux',
     'datadog.telemetry.Uname.EmptyUtsName',
     'datadog.telemetry.dependency.LocationsCollectingTransformer',
+    'datadog.telemetry.dependency.JbossVirtualFileHelper',
     'datadog.telemetry.TelemetrySystem',
     'datadog.telemetry.api.*'
   ]
